% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanFormat.R
\name{cleanData}
\alias{cleanData}
\title{Clean a single Ecobici csv file}
\usage{
cleanData(csvdata)
}
\arguments{
\item{Ecobici}{data frame of an Ecobici csv file.}
}
\value{
a data frame with clean names and columns in the right order.
}
\description{
This function will take a csv file and apply the correct format to each column.
}
