% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanFormat.R
\name{CleanFormat}
\alias{CleanFormat}
\title{Clean Ecobici csv files}
\usage{
CleanFormat(Ecobici)
}
\arguments{
\item{Ecobici}{list containing all Ecobici csv files.}
}
\value{
an object of class 'Ecobici' with a data frame containing the joined and cleaned version of the Ecobici files.
}
\description{
This function will take the uploaded csv files, make the corresponding data cleaning to then join them in a single table.
}
