% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTables.R
\name{CreateTables}
\alias{CreateTables}
\title{Create summary tables}
\usage{
CreateTables(Ecobici)
}
\arguments{
\item{Ecobici}{list containing all Ecobici information.}
}
\value{
an object of class 'Ecobici' with an additional element containing the summary tables.
}
\description{
This function will create the summary tables necessary for the report.
}
