% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{Transform Ecobici variables}
\usage{
TransformData(Ecobici)
}
\arguments{
\item{Ecobici}{list containing all Ecobici information.}
}
\value{
an object of class 'Ecobici' with a new variable for the length of the trip.
}
\description{
This function will transform and create new variables from Ecobici data.
}
