% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{ValidateInput}
\alias{ValidateInput}
\title{Ecobici Input Data Validation}
\usage{
ValidateInput(Ecobici)
}
\arguments{
\item{Ecobici}{list containing all Ecobici information.}
}
\value{
an object of class 'Ecobici' with a new table called validation indicating which conditions were met.
}
\description{
This function will validate data from Ecobici data that has been uploaded and transformed
previously. There are specific conditions that the data should have to continue with the next
steps of the analysis.
}
