% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\name{ReadFiles}
\alias{ReadFiles}
\title{Read Ecobici csv files}
\usage{
ReadFiles(input_path = NULL)
}
\arguments{
\item{input_path}{character string of the location of the files.}
}
\value{
an object of class 'Ecobici'.
}
\description{
This function will read all csv files from an specified folder.
}
